mgr: pass python interpreter's path to embedded python
* also do prevent python from registering its own signal handler, it does
  not make sense in our embedded use case.
* pass python interpreter's path to embedded python before initializing
  it. python uses this path to look up the "site" modules, etc.

so it can use other interpreter if it is not intalled into $PATH,
otherwise the "python" in $PATH will always be used even the
PYTHON_EXECUTABLE in CMake's cache is pointing to another python
interpreter.

if one want to debug ceph-mgr with a customize build python, this would
be helpful.

Signed-off-by: Kefu Chai <kchai@redhat.com>