crush/CrushWrapper: remap across intervening bucket types
The previous code could only swap overfull devices with underfull devices
if they were in the same bucket.  With this change, we can map across
buckets.  For example, if host A has more PGs than host B, then we'll
remap some PGs on devices in host A with devices in host B.

Signed-off-by: Sage Weil <sage@redhat.com>