logclient: use the seq id of the 1st log entry when resetting session
because do_log() does not update last_log (see LogClient::get_next_seq())
and log_queue atomically, and log_queue is updated *after* last_log, so
we can not use them to figure out the number of log entries to be sent.
so, we choose to trust log_queue in this change.

Fixes: http://tracker.ceph.com/issues/19427
Signed-off-by: Kefu Chai <kchai@redhat.com>