ceph-disk: separate ceph-osd --check-needs-* logs
It is using the OSD id zero but have nothing to do with OSD zero and
this is confusing to the user. The log themselves do not need to be kept
around and are stored in the run directory so that they can be disposed
of after reboot.

Fixes: http://tracker.ceph.com/issues/19888

Signed-off-by: Loic Dachary <loic@dachary.org>
(cherry picked from commit c7b3c46bd63b78475868e405bf20d9c142f0336a)