mon: Bring daemonization in line with other daemons
ceph-mon doesn't use the global_init_daemonize() as the other
daemons do.  This resulted in the daemon staying in the session
of the process that started it.  In testing if I terminated a
test script only ceph-mon would be exit.

Signed-off-by: David Zafman <dzafman@redhat.com>
(cherry picked from commit 19da3e8441a5055a2946c88eabb7ddf14c92c144)