ReplicatedPG: add CHECKSUM->CMPEXT req translation
With the transition to the cmpext upstream API, we need to continue to
handle iSCSI gw clients that use the old API. Do so via translation.

This logic should be removed once clients using the old API are no
longer supported, as was done with 8047d61183 ("ReplicatedPG: drop
CACHE_PIN->WRITESAME req translation"). (bsc#1035937).

Signed-off-by: David Disseldorp <ddiss@suse.de>