ceph-detect-init: detect init system by poking the system not checking its codename
ubuntu's codename is not strictly lexical ordered. and after Xenial,
Yakkity and Zesty, it will be Artful. so we can not just compare it with
'vivid' to tell if it is using systemd or upstart.

also, the current approach for debian only checks for squeeze and
wheezy. this is not flexible and per BTS#862075, it's wrong. so, we'd
better check the system by poking it instead of just looking at its
codename

---
[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=862075

Fixes: http://tracker.ceph.com/issues/19884
Signed-off-by: Kefu Chai <kchai@redhat.com>