ceph-disk: remove unused methods
this is a follow-up of aac89719881c2788941b74d385d95860e520ea78, which
replaced partx with partprobe. and we don't need to detect the platform
to decide if we need to use partx or not anymore with that commit
because we are dropping the support of distros w/o partprobe.

Fixes: http://tracker.ceph.com/issues/19884
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit b7c0949f361cd23c04efc3bcc5fe4bbf247e5d68)

Conflicts:
    src/ceph-disk/ceph_disk/main.py (trivial resolution)