rgw: add more information to the "req done" log message
As discussed in [1], add some further information when logging a finished
request in rgw:

- op_name
- user_id
- domain
- request_uri
- request_params
- obj_size
- duration

With the duration being show here, we do not need the request start
being logged, move it to a higher level.

In the end it would be nice to aggregate all this information together
with the additional data that civetweb is keeping, in particular
remote_addr and num_bytes_sent, but this seems useful as a first step.

[1] http://tracker.ceph.com/issues/19917

Signed-off-by: Jens Rosenboom <j.rosenboom@x-ion.de>