osd/PrimaryLogPG: fix oi reset during trim_object
We partially fixed this in 5d95cc9861aa4b8d69b680b96a32626eccf9fab8
but cleared the wrong oi (the clone's and not the
snap/head object's).

While we're here, drop the various bits where we clear
state bits; these are about to get clobbered when we reset
oi anyway.

Fixes: http://tracker.ceph.com/issues/19947
Signed-off-by: Sage Weil <sage@redhat.com>