qa/workunits/rados/test-upgrade-*: whitelist tests for master
The jewel-x upgrade test now runs this script against a mixed cluster on
a machine with code from master installed.  That means we have to
skip any new tests that will fail on a mixed cluster. CloneRange was
removed in 0d7b0b7.

Signed-off-by: Kefu Chai <kchai@redhat.com>
Conflicts:
    qa/workunits/rados/test-upgrade-v11.0.0.sh: this change is not
cherry-picked from master, because the clone-range op was removed from
master. and only supported in pre-luminous releases.