mgr: expose a MgrMap in PyModules
Previously mgrmap was only handled at the
MgrStandby level, to decide when to go active.

Some modules might be interested in the map so
that they can tell the user about the mgr
configuration.

Pass a copy upwards through the layers so that
PyModules doesn't get into locking hell trying to
reach down into Mgr/MgrStandby to grab it.

Signed-off-by: John Spray <john.spray@redhat.com>