mon/OSDMonitor: add osd_pool_default_expected_objects_per_pg option
By default, set the expected_num_objects for a pool to be enough objects
per PG so that FileStore will do some of its initial splitting.  This
avoids a throughput collapse later when the pool starts to fill and we
have to do a split.  It also avoids moving objects around between
directories and subverting XFS attempts to keep files in the same directory
within the same AG.

Initially set the default to 0 here; we'll crank it up later once we have
more testing.

Signed-off-by: Sage Weil <sage@redhat.com>