tests: upgrade/hammer-x/v0-94-6-mon-overload: tweak packages list
Include some hammer dependencies that aren't in the jewel default packages
list, and exclude some java packages that may not be in the hammer repo and are
not needed for the upgrade test in any case.

N.B.: This cannot be cherry-picked from master because upgrade/hammer-x was
dropped in master.

Signed-off-by: Nathan Cutler <ncutler@suse.com>