Fix s3 object uploads with chunked transfers and v4 signatures.
With aws-sdk-java 1.11, large uploads use chunked transfer by default,
and v4 signatures are the default.  The java sdk uses a slightly different
string "AWS4-HMAC-SHA256-PAYLOAD" when constructing the per-chunk signature
than ceph was using.  This same string also appears in a current
copy of s3-api.pdf , so it must be the more correct value.

Fixes: http://tracker.ceph.com/issues/20447
Signed-off-by: Marcus Watts <mwatts@redhat.com>