rgw: add tail tag to track tail instance
Use tail_tag instead of idtag. The tail_tag is modified when either
creating a new tail, or adding a refcount to the tail (e.g., when
copying object). When just modifying the object's head, tail_tag
should remain the same. The gc will then use the tail_tag (if
exists, idtag if not) as the refcount id.

Fixes: http://tracker.ceph.com/issues/20234

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>