RemoteProcess: command can be now executed in specified directory.
RemoteProcess, a class built around Paramiko's exec_command(),
currently lacks support for the "cwd" parameter. It specifies
the directory in which the requested command should be executed.
This stays in contrast to LocalRemoteProcess from the vstart
runner as it wraps Python's subprocess, and thus naturally
offers the feature. The patch unifies these two implementations.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>