mds: revise health checks for _FAILED/_DEGRADED
No longer output MDS_* versions as well as FS_* versions,
because it was noisy and the important message is about
the availability (or not) of the filesystem.

Revise the _FAILED check to only raise the message if
there are not suitable replacements available for failed
ranks.  This avoids a spurious health check failure when
a rank has been failed (e.g. by the admin) but it will
be replaced at the next tick().

After this change, doing a "ceph mds fail" when a standby
is available just gives you a single FS_DEGRADED health
check from the point of the "fail" to when the replacement
is active.

Signed-off-by: John Spray <john.spray@redhat.com>