librados: 'pg ls' is not a mgr command
Note that this breaks the command *during* the mon upgrade from jewel ->
luminous, which is slightly annoying, but means we avoid proxying via the
mon after upgrade is complete, which is good and IMO more important.

In the future we may want librados to cache the command descriptions so
that commands can be directed automatically.

Signed-off-by: Sage Weil <sage@redhat.com>