osd: do not send empty MOSDFull msg
the osd stat is updated by the beartbeat thread, while
OSD::send_full_update() is normally called in the
tick_timer_without_osd_lock's thread. so there is rare chance that the
full state is reset after `service.need_fullness_update()` returns true.
this does not hurt at all. but it confuses the mon by adding its burden.
so avoid this if possible.

Signed-off-by: Kefu Chai <kchai@redhat.com>