use STL find() and iterators instead of count()
This PR is 1st part of replacement of if(container.count(foo))return container[foo]; with container.find(foo). Since the .count(foo) code searches for key till end and returns TRUE/FALSE, it's good to quit the search if we find the key at 1st instance.
After this PR gets rolled out, similar PRs would be created for other modules as well.

Resolves: http://tracker.ceph.com/issues/432

Signed-off-by: Amit Kumar <amitkuma@redhat.com>