osdc/Objecter: unify disparate EAGAIN handling paths into one
We had two different code paths for EAGAIN that both
seemed to resend the same request again (with a new
tid), but implemneted it in two different ways. Further,
the first one was conditioned on the BALANCE_READS
flag but not LOCALIZED_READS, which should (I think) be
equivalent in this context.

Unify them into a single path.

Signed-off-by: Sage Weil <sage@redhat.com>