osd: do not send pgstats unless luminous
MgrClient send_report() upon MMgrConfigure when it is starting. and
send_report() sends a pgstats to mgr. but osd is not supposed to update
mgr with pgstats unless luminous is required. if we send a pgstats, the
stale pgstats will be kept by mgr until "require-osd-release luminous".

this false alarm is annoying and misleading to those who are upgrading
to luminous.

Signed-off-by: Kefu Chai <kchai@redhat.com>